requirements.txt
Python == 3.8.5
Django == 4.2.15
djangorestframework == 3.15.2
psycopg2 == 2.9.9
Faker == 27.0.0
